move scrape functionality to a seperate file and create structure for CLI
